AI-agent-pocs/my-mutual-fund-agent/requirements.txt
langchain  
langchain-community 
langchain-anthropic
langchain_google_genai
python-dotenv
pydantic